requirements.txt
Django==3.2.9
django-allauth==0.46.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-js-asset==1.2.2
django-mptt==0.13.4
django_tables2==2.4.1
django-widget-tweaks==1.4.9
Pillow==8.4.0
psycopg2==2.9.2
